Feat: created game class
